server/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
google-auth==2.6.0
requests==2.28.1
